Move ViewService from Reko.UserInterfaces.WindowsForms to Reko.Gui
Move TextFileViewerServiceImpl and ViewService from
Reko.UserInterfaces.WindowsForms to Reko.Gui. They are not
Windows Forms specific
